requirements.txt
certifi==2022.9.14
cycler==0.11.0
distlib==0.3.6
filelock==3.8.0
fonttools==4.33.3
kiwisolver==1.4.3
matplotlib==3.5.2
mpmath==1.2.1
numpy==1.22.4
packaging==21.3
pandas==1.4.4
Pillow==9.1.1
pipenv==2022.9.21
platformdirs==2.5.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
six==1.16.0
sympy==1.10.1
virtualenv==20.16.5
virtualenv-clone==0.5.7
